s2_hackatones-back/src/test/java/co/edu/uniandes/csw/hackatones/test/persistence/LugarPersistenceTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package co.edu.uniandes.csw.hackatones.test.persistence;
import co.edu.uniandes.csw.hackatones.entities.LugarEntity;
import co.edu.uniandes.csw.hackatones.persistence.LugarPersistence;
import javax.inject.Inject;
import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import org.jboss.arquillian.container.test.api.Deployment;
import org.jboss.arquillian.junit.Arquillian;
import org.jboss.shrinkwrap.api.ShrinkWrap;
import org.jboss.shrinkwrap.api.spec.JavaArchive;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import uk.co.jemos.podam.api.PodamFactory;
import uk.co.jemos.podam.api.PodamFactoryImpl;



/**
 *
 * @author jd.monsalve
 */
@RunWith(Arquillian.class)
public class LugarPersistenceTest {
    
    @Deployment
    public static JavaArchive createDeployment(){
        return ShrinkWrap.create(JavaArchive.class)
                .addClass(LugarEntity.class)
                .addClass(LugarPersistence.class)
                .addAsManifestResource("META-INF/persistence.xml","persistence.xml")
                .addAsManifestResource("META-INF/beans.xml","beans.xml");
    }
    
    @Inject
    LugarPersistence cp;
    
    @PersistenceContext
        EntityManager em;
    
    @Test
    public void createTest(){
        PodamFactory factory = new PodamFactoryImpl();
        LugarEntity lugar = factory.manufacturePojo(LugarEntity.class);
        LugarEntity result = cp.create(lugar);
        Assert.assertNotNull(result);      
        
        LugarEntity entity = em.find(LugarEntity.class,result.getId());
        Assert.assertEquals(lugar.getNombre(), entity.getNombre());
    }
    
    
}
